put switch button under welcome page to keep it consistent
