+++ src/renderer/components/app-load-fallback/index.tsx
@@ -4,7 +4,7 @@ import { LocaleContext } from '../../contexts';
 import { noOp } from '../../env-and-consts';
 import detectLanguage from '../../services/i18n/simple-language-detector';
 import { SupportedLocale } from '../../types';
-import { StyleSwitcher } from '../style-switcher';
+import StyleSwitcher from '../style-switcher';
 
 const textTranslations = {
   [SupportedLocale.ZH_CN]: '载入中...',
