+++ src/renderer/components/welcome-page/index.tsx
@@ -5,9 +5,9 @@ import Row from 'react-bootstrap/Row';
 import { useTranslation } from 'react-i18next';
 
 import { Positional } from '../helper-components';
-import { StyleButton } from '../style-switcher';
 
-import startButton from './start-button.svg';
+import startButton from './start-button-img.svg';
+import StyleButton from './style-button';
 
 const WelcomePage: FC<{ onNext: MouseEventHandler<HTMLElement> }> = ({ onNext }) => {
   const { t } = useTranslation();
